// src/app/core/models/penguin.ts
export interface Penguin {
  id?: number;
  tokenURI?: string;
  tokenId?: string;
  isMinted?: boolean;
  price?: number;
  numberOfTransfers?: number;
  isForSale?: boolean;
  mintedBy?: string;
  currentOwner?: string;
  background: string;
  skin: string;
  eye: string;
  mouth: string;
  cloth: string;
  glass: string;
  hair: string;
  hat: string;
  neck: string;
  random_object: string;
}